Visual Studio 2008: Add freetype Projects
This adds a set of Visual Studio 2008 project files that is used to build
freetype, along with a customized ftoption.h (which will be used during
the build and "installed" after the completion of the build) that is used
to build freetype as a DLL and enables the optional features which enables
bzip2-compressed file support and PNG bitmap support, along with using the
system-supplied zlib for gzip-compressed file support, since we already use
zlib in many parts of the GTK+/Clutter stack.

This means that this freetype DLL will depend on the libbz2, zlib and
libpng (current 1.6.x) DLLs.

This builds the freetype in both x86 and x64 flavors, used as an optional
dependency of cairo, and also in PangoFT2.