Add Projects to build libiconv
This adds MSVC 2008-2013 Project Files to build libiconv-1.14, including
the libcharset and libiconv DLLs/LIBs and the iconv utility.  Unlike the
other projects that other project file sets build, this is only usable for
building libiconv-1.14 (1.14 is the latest stable release of libiconv
for now, though), due to the intensive use of gnulib that gets updated
between releases, rendering project files and source updates out-of-date.