gettext: Add Patches To Enable Sources To Build
..the latest release, 0.19.4, with Visual Studio.

This is one hard part to make the GTK+ stack (and also for KDE, for
references, as it also uses gettext) buildable under Visual Studio.

Gettext uses quite a bit of gnulib items that need proper configuration
so that intl.dll and the tools especially, can be built fully with Visual
Studio.  At this point though, most of the sources can be built with Visual
Studio 2008+ out of the box, except:

-As stated, the config.h.in and the *.in.h headers must be configured
 first, this is the bulk of the patches.
-For gettext-tools, the headers also need to be adapted to export/import
 symbols and public variables using __declspec(dllexport) and
 __declspec(dllimport), for both generated and real source headers, also a
 significant part of the patches.
-A few sources needed updates, mainly to ensure variables are declared in
 the right spots for C89 and to include needed headers.
-A few headers and generated files need to be placed in the right places
 beforehand, and need to be adapted so that the necessary symbols can be
 exported properly.
-xgettext.c needs to be built in C++ mode, so some header inclusion in
 there must be done in C-mode (i.e. extern "C"), so that the symbols can
 be properly imported during linking.
-configmake.h needs to not included unconditionally, and we need to not
 define fcntl as rpl_fcntl.

Note: The gnulib headers are filled with gnulib-specific macros, but they
do not get into the way as one can force-include the headers in
build-aux/snippet, in which gnulib already had consideration for correctly
building those under Visual Studio, hence we are able to build the gnulib
items without much issue provided the *.in.h and config.h.in are properly
configured.  This also means that a new set of patches will be needed for
each gettext release as it will pull in new gnulib items with every
release :|-meaning diff'ing big time with each release!