ICU: Add Initial Visual Studio 2008 Projects
This adds a set of Visual Studio 2008 Projects which will allow one to
build ICU 55.x+ without the use of Cygwin, in both Win32 and x64 flavors.

This will require a patch to the NMake Makefiles for generating the actual
data DLL (the final icudtxx.dll), and does not include the "Installation"
portion.

This will be used for libxml2.