ICU: Complete MSVC 2008 Projects
Add the install project that will copy the build results, and tweak the
"install" path so that it will fit better with the other projects in the
stack.

One should be aware that iculeXX.dll and iculxXX.dll is considered
deprecated and should consider using HarfBuzz-ng for layout purposes, and
it seems that iculeXX.dll and iculxXX.dll does not work well for x64
configs, in addition to the many open tickets in upstream reported against
it and it was not in active development for quite a while, according to
the ICU documentation.

Also fix the icudt project so that it creates the directories as a
post-build event, to avoid errors when cleaning.