libxml2: Add Pre-configured Headers
These are the build-time headers that are normally generated by the
win32/configure.js script, which is used in our projects to proceed with
the build of libxml2.

The projects also include a custom build step, which is disabled by
default, which also triggers the configure.js script, as it is always
re-run, causing the libxml2 build to re-start.