PCRE: Add MSVC Projects to Build PCRE
This adds Visual Studio 2008-2013 project files that can be used to build
PCRE, which was done against the latest 1.x stable release, 8.37.  Attempts
to add PCRE 10.x (2.x) will be done when GLib/QT can work against PCRE 2.x.

This is becoming necessary as there are plans in this dev cylcle that GLib
will not bundle its patched PCRE sources due to difficulty to keep it
up-to-date with certain thing, which means that GLib will most probably
hard-depend on an external PCRE installation for GRegex.

This is also done to help prepare one to be able to be ready to have the
dependencies to build QT without using its bundled sources for those deps,
which can add size to the built DLLs/programs.

This also contains the pre-configured build-time headers, which are adapted
from the ones generated by CMake, to be usable across different Visual
Studio versions.