pixman: Add Patch To Enable Threads Tests
The original pixman thread test uses pthreads, which, although works in
Windows, but would then require that we install a Windows implementation of
pthreads.

Instead of requiring everyone to do this, add a patch that would update
pixman's thread test to use Windows native threads when built on Windows,
so we can know whether our build of pixman is thread-safe (which, in fact,
it is, as mentioned in the bug report stated in the patch).