pixman: Add Visual Studio Projects
This builds pixman as a DLL, with all the test programs, where the tests
are run as they are being built, before the .dll/.lib and headers are
copied, in both 32-bit and 64-bit configurations.

NB: Due to an internal compiler error, Visual Studio 2008 x64 builds do not
include support for the SSSE3 instruction set; 2010 and above do include
such support for all builds.  Since MMX intrinistics are not support by x64
builds, according to Microsoft, x64 builds do not include MMX support.