graphite2: Add Visual Studio Projects
graphite2 is normally built by using CMake, but we simplify the process by
just using Project files to build the graphite2 DLL, which does not go
through many configuration processes.  This does not include the tests that
the CMake builds use, but running the test programs seems to be fine on
both x64 and x86 builds.