HarfBuzz: Add Visual Studio Projects
This will build the main Harfbuzz DLL, and the ICU and GObject extension
DLLs/Libs for Harfbuzz, in both x86 and x64 configurations, with the tools
in util/ and the tests in test/api.  The tests are run as part of the
build, and the main Harfbuzz include all features relevant to Windows
except fontconfig.