vs9/allinone.sln: Build GDK-Pixbuf
This will include GDK-Pixbuf into the build process of the all-in-one
build.  Note that this builds the flavor using the opensource imaging
libraries (libpng, libjpeg-turbo, libtiff, libjasper), as this is in a
better shape than the existing GDI+ image loaders.  All of the image
loaders shipped with GDK-Pixbuf are built into the main GDK-Pixbuf DLL.