gtk2: Add Patch To Update Project GUID
Add a patch to update the GUID for the gtk-install projects so that it will
not conflict with the pango-install project.  This patch is done upstream
as well, but since we might not have GTK+-2.24.x releases that often, add
this patch here anyways, until the next GTK+-2.24.x release takes place.