Add Libcroco Sources
This checks in the entire libcroco package, which is from the tarball that
is done with 'make dist' from the latest master git checkout, as there was
not any libcroco releases lately and a number of improvements have went in.

This also includes a change in the value of GlibEtcInstallRoot to
build/win32/vsX/croco-version-paths.vsprops|props so that it would fit
better with our all-in-one solution layout; otherwise everything else are
left untouched from a tarball that is built from the latest libcroco git
checkout.

If a libcroco release does come to pass, everything here besides the patch
to patch build/win32/vsX/croco-version-paths.vsprops|props will be removed.