gettext: Allow build for 0.19.7 on Visual Studio
This adds the patches to add the pre-configured and generated sources so
that the code in gettext-runtime and gettext-tools can be built using
Visual Studio 2008~2015, as well as source updates that are necessary
during the process.

This also adds .def files for the gettextsrc and gettextlib DLLs so that
we can minimize the source modifications that are necessary.