libiconv: Add pre-configured config.h and gnulib headers for 1.15
This adds the headers from their .h.in and in.h counternparts that are
generated through the official MSVC build instructions via Cygwin and
then merged for the different x86/x64 configs of the various Visual
Studio versions supported (2008~2015).  The libcharset and libiconv
DLLs are supported here, as well as the iconv tool, with or without
gettext support.

These do not replace any files in the libiconv-1.15 source tree and are
looked for first by the projects so that the proper configs are
acquired.