libjpeg-turbo: Update projects for 1.5.x
Adapt the projects to build 1.5.x, and move the projects to win32/ from
build/win32/, so that we can go down one less layer in the source tree
hierarchy.  Also simplify the test .bat script for [c|d]jpeg.exe as the
sources now bundle an md5 checking tool that we now build and use to
test the build.

Tested with 1.5.1.