tiff: Move projects to win32/ from build/win32/
This makes it more in-line with the cases in the other parts of the
stack.

Also add projects for Visual Studio 2017, and support LZMA compression
on Visual Studio 2013 and later using liblzma from XZ utils.

Remove projects for gif2tiff, gmp2tiff and ras2tiff as they are not
maintained and are removed from the distribution.