Add NMake Build system for fontconfig
This adds a default NMake-based build system for building the latest
stable fontconfig using Visual Studio, where:

-Choice of using Expat or libxml2 is supported.
-The default Windows directories are used.
-The generated headers needed for the build are all re-generated during
 the build.  Note that since the fcalias*.h and fcftalias*.h are used
 for GCC builds only, they are generated as stubs, as they are not
 meaningful on Visual Studio builds.  We export symbols frmo the
 fontconfig DLL by defining FcPublic to be __declspec(dllexport) during
 the build of the DLL.

For pre-Visual Studio 2015 builds, the snprintf() implementation from
libtiff is used, along with a public domain implementation of dirent
functions for all Visual Studio builds, which is the same as the one
that is shipped with GLib.

Note that there are no code changes in the vanilla FontConfig sources.