ICU: Add NMake Makefiles to build ICU4C 59.x
This builds the ICU4C 59.x DLLs and tools, with the data built as a DLL
using NMake.  Due to code requirements, Visual Studio 2015 or later will
be required.

The source/data/makedata.mak and source/test/testdata/testdata.mak are
drop-in replacements for the ones in the sources, and can be used either
with this set of the NMake Makefiles, or with the 2015 projects that are
provided with the sources.