libpng: Add NMake Makefiles
It turns out that the Visual Studio projects get cumbersome to maintain,
so add NMake Makefiles that can be used to build the libpng DLL with the
tools and the tests, and running the tests as they were in the project
files.

The projects will remain for now, but they will be superseded with the
NMake Makefile set.