zlib: Add NMake Makefiles
As it turns out to be quite tedious to maintain the MSVC projects for
each MSVC version, add NMake Makefiles that can be used across MSVC
versions to build ZLib.

We still provide an option to use the MASM optimizations, but there are
stuff in place in the code to warn against that (and it currently cause
crashes).