Add Nake Makefiles to build sqlite3
This adds NMake Makefiles to build the sqlite3 DLL and the sqlite3 tool,
as well as the TEA/TCL bindings for sqlite3, if desired.  This also
enables ICU support in the built sqlite3 DLL, if desired.

Note that the .def file from the pre-built DLL package is *not* required
for this, as we defined macros to export symbols from the built DLL.

This can be used on any of the amalgamation source releases of sqlite3;
do note that the TEA/TCL bindings can be built only with the
amalgamation source tarballs that contain the autotools scripts and the
TEA/TCL sources, and TCL needs to be built and installed beforehand.