sqlite3: NMake-Split out TCL portion in its own NMake Makefile
The proper way to build the tcl extension for sqlite is to have the DLL
named as sqlite3<sqlite3 version without dots>t.dll.  To do that, we split
out the portion where we build the TCL extension out into its own NMake
Makefile.

This is so that we can first build and run nmakehlp.exe, and using its
results to call the new Makefile so that we can properly pass in the
required PACKAGE_VERSION, and name the final DLL with the sqlite3
version info.