gettext/0.20.1: Build the tools in gettext-tools
This adds to the NMake Makefiles to build the tools in
gettext-tools/src, which will also in turn build libgettextsrc and
libgettextlib as DLLs.  Note that the DLLs for libgettextsrc and
libgettextsrc is versioned, as the APIs in there is only valid for this
release.

Also build libgettextpo as a DLL as well.