gettext: Add pre-configured headers for gettext-0.21
This prepares for building with Visual Studio using NMake.  The headers
have been generated using a Cygwin build that is set up for Visual
Studio builds, so that the proper compiler settings are obtained for
these.

gnulib can be quite a pain since we need to massage autotools with the
right arguments...