gettext: Add NMake Makefiles for 0.21
This adds a set of NMake Makefiles that can be used to build gettext
with Visual Studio 2008 and later, for i686 (x86/Win32), x86-64 (x64)
and aarch64 (arm64), with the gettext runtimes (intl.dll, asprintf.dll),
the libtextstyle and libgettextpo DLLs, as well as the utility programs
provided with the package, plus the C# version of the gettext library.

Facilities to copy the build results to a folder is not yet added, but
we can be assured that it is now possible to build gettext with the
only the visual Studio tools (plus libiconv) installed.