gettext/0.21: Add build option for libtool-style DLL names
For people who may prefer to use the DLLs (libintl, libgettextpo,
libasprintf and libtextstyle) built here to be drop-in replacements for
the MinGW-built ones that are more prevalent in their MSVC
toolchain/collection of libraries, a build option is added to the NMake
Makefiles to make the above-named DLLs named like those built with the
autotools build system.

Note that this is not applied to libgettextsrc and libgettextlib, as
these DLLs are not meant to be used as a library with stable API, but is
more for the tools to make use of, so they continue to be named with the
version of the current gettext version.