gettext: Fix locating rule search
The original code to look for locating/ITS rules only works if dirent is
enabled for the build, which, incidentally is not enabled on Visual
Studio builds as Visual Studio does not ship with a dirent.h (doh!).

So, since we are already building the dirent items in gnulib-lib in
gettext-tools, just add a patch for locating-rules.c to always include
dirent.h (either from the system or from gnulib), and make sure that our
gettextlib DLL exports the symbols needed for dirent support for our
builds.

Like in its.c, also make sure that we dllimport xmlFree() correctly.