libffi/3.4.3: Add NMake Makefiles for building
This set of NMake Makefiles support building libffi as a DLL for Visual
Studio 2008 and later, for x86 (32-bit and amd64) and ARM64 builds, in
conjunction with the appropriate patches in the previous commit.  Only a
standard installation of Visual Studio with the proper build tools for
the hardware architecture is required for this to work.

This will be what I think would be the easiest way to build libffi
before CMake or Meson support lands in the upstream sources.