gettext: Add preliminary support for building 0.21.1
This adds support to use the NMake Makefiles to build gettext-0.21.1,
for gettext-runtime and libtextstyle for now.  The support for building
the gettext-tools items are not ready at this point, so should not yet
be used, and will be done later.

Note that from this release of gettext, only Visual Studio 2015 or later
is supported, so Visual Studio 2013 and earlier will need to make do for
gettext-0.21.