gettext/0.21.1: Add build files for gettext-tools
This cmopletes our build suite to build libgettextpo and the gettext
tools (that are installed) in gettext-tools.

This support is still WIP as no install target is provided at the
moment, which will be added later, but the full set of tools and
libraries are now buildable for 0.21.1 with this NMake Makefile set.